extracting the logic for widget extraction into its own files
