+++ src/tui/browser.rs
@@ -1,15 +1,12 @@
-use crate::database::connect::add_history;
 use crate::search::link::Link;
 use crate::search::scrape::format_url;
-use crate::transform::cache::{get_content, preload};
 use crate::transform::page::PageExtractor;
 use crate::tui::browser::Action::{Down, Exit, Next, Open, PageDown, PageUp, Previous, Up};
-use crate::tui::display::{default_block, Display, TUI_BORDER_COLOR};
+use crate::tui::browser_widgets::{draw_page_numbers, new_page};
+use crate::tui::display::Display;
+use crate::tui::general_widgets::default_block;
 use crossterm::event;
 use crossterm::event::{KeyCode, KeyEvent, KeyEventKind, KeyModifiers};
-use ratatui::layout::Alignment;
-use ratatui::prelude::{Color, Line, Style, Text};
-use ratatui::widgets::{Paragraph, Wrap};
 
 static PAGE_INSTRUCTIONS: &str = " Quit: q/Esc | Scroll Down: j/↓ | Scroll Up: k/↑ | Page Down: CTRL+d | Page Up: CTRL+u | Next: n/→ | Back: b/← | Open in Browser: o ";
 
@@ -138,51 +135,4 @@ fn open_link(index: usize, links: &[Link]) {
         .and_then(|url| open::that(&url).err())
         .iter()
         .for_each(|e| println!("{e}"));
-}
-
-pub fn new_page(
-    index: usize,
-    links: &[Link],
-    extractor: &PageExtractor,
-    history_active: bool,
-) -> (String, Paragraph<'static>) {
-    if let Some(link) = links.get(index + 1) {
-        preload(link, extractor); // Initiate the call to get the page after this one
-    }
-    links
-        .get(index)
-        .inspect(|link| {
-            if history_active {
-                _ = add_history(link);
-            }
-        })
-        .map(|link| (link, get_content(link, extractor)))
-        .map_or_else(
-            || {
-                (
-                    String::from("None"),
-                    Paragraph::new(Text::from(String::from("Index out of bounds"))),
-                )
-            },
-            |(link, paragraph)| {
-                (
-                    extract_title(link),
-                    paragraph
-                        .style(Style::default().fg(Color::White))
-                        .wrap(Wrap { trim: false })
-                        .scroll((0, 0)),
-                )
-            },
-        )
-}
-fn extract_title(link: &Link) -> String {
-    format!(" {} ({}) ", link.title, link.url)
-}
-
-pub fn draw_page_numbers(index: usize, pages: usize) -> Text<'static> {
-    Text::from(Line::styled(
-        format!(" [{index}/{pages}] "),
-        *TUI_BORDER_COLOR,
-    ))
-    .alignment(Alignment::Right)
 }
