+++ src/tui/history.rs
@@ -1,28 +1,16 @@
 use crate::actions::direct;
-use crate::config::load::Config as LocalConfig;
 use crate::database::connect::{remove_history, HistoryData};
-use crate::tui::display::{default_block, Display};
+use crate::tui::display::Display;
+use crate::tui::general_widgets::default_block;
 use crate::tui::history::Action::{Backspace, ChangeSearch, Continue, Down, Exit, Open, Text, Up};
 use crate::tui::history::SearchOn::{Title, Url};
-use chrono::{NaiveDateTime, Utc};
+use crate::tui::history_search::order_by_match;
+use crate::tui::history_widgets::{create_table, draw_history_count, draw_search_text};
 use crossterm::event;
 use crossterm::event::{KeyCode, KeyEvent, KeyEventKind};
-use nucleo_matcher::pattern::{AtomKind, CaseMatching, Normalization, Pattern};
-use nucleo_matcher::{Config, Matcher, Utf32Str};
-use once_cell::sync::Lazy;
-use ratatui::layout::{Alignment, Constraint};
-use ratatui::prelude::Modifier;
-use ratatui::style::{Color, Style};
-use ratatui::text::{Line, Span};
-use ratatui::widgets::{Cell, Paragraph, Row, Table, TableState};
-use std::cmp::Ordering;
+use ratatui::widgets::TableState;
 use Action::Delete;
 
-static URL_COLOR: Lazy<Style> = Lazy::new(LocalConfig::get_url_color);
-static TITLE_COLOR: Lazy<Style> = Lazy::new(LocalConfig::get_title_color);
-static TIME_COLOR: Lazy<Style> = Lazy::new(LocalConfig::get_time_color);
-static SEARCH_TYPE: Lazy<AtomKind> = Lazy::new(LocalConfig::get_search_type);
-static TEXT_COLOR: Lazy<Style> = Lazy::new(crate::config::load::Config::get_text_color);
 static HISTORY_INSTRUCTIONS: &str =
     " Quit: Esc | Scroll Down: ↓ | Scroll Up: ↑ | Open: ↵ | Delete: Delete | Tab: Change search ";
 
@@ -49,8 +37,7 @@ impl History {
         let mut search_on = Title;
         history = order_by_match(&mut history, &mut user_search, &search_on);
         state.select_last();
-        let mut rows = create_rows(&history, &user_search, &search_on);
-        let mut table = create_table(&mut rows);
+        let mut table = create_table(&history, &user_search, &search_on);
         let border = default_block(" History ", HISTORY_INSTRUCTIONS);
         let mut search = draw_search_text(&user_search, &search_on);
         let mut entry_count = draw_history_count(history.len() as u16);
@@ -104,7 +91,7 @@ impl History {
                     let removed = history.remove(state.selected().unwrap_or(0));
                     _ = remove_history(&removed.url);
                     total_history.retain(|item| *item != removed);
-                    table = create_table(&mut create_rows(&history, &user_search, &search_on));
+                    table = create_table(&history, &user_search, &search_on);
                     *state.offset_mut() = state.offset().saturating_sub(1);
                     entry_count = draw_history_count(history.len() as u16);
                     self.display.draw_history(
@@ -119,7 +106,7 @@ impl History {
                 Text(char) => {
                     user_search.push(char);
                     history = order_by_match(&mut history, &mut user_search, &search_on);
-                    table = create_table(&mut create_rows(&history, &user_search, &search_on));
+                    table = create_table(&history, &user_search, &search_on);
                     *state.offset_mut() = 0;
                     state.select_last();
                     search = draw_search_text(&user_search, &search_on);
@@ -137,7 +124,7 @@ impl History {
                     user_search.pop();
                     history.clone_from(&total_history);
                     history = order_by_match(&mut history, &mut user_search, &search_on);
-                    table = create_table(&mut create_rows(&history, &user_search, &search_on));
+                    table = create_table(&history, &user_search, &search_on);
                     *state.offset_mut() = 0;
                     state.select_last();
                     search = draw_search_text(&user_search, &search_on);
@@ -155,7 +142,7 @@ impl History {
                     search_on = next_search(&search_on);
                     history.clone_from(&total_history);
                     history = order_by_match(&mut history, &mut user_search, &search_on);
-                    table = create_table(&mut create_rows(&history, &user_search, &search_on));
+                    table = create_table(&history, &user_search, &search_on);
                     *state.offset_mut() = 0;
                     state.select_last();
                     search = draw_search_text(&user_search, &search_on);
@@ -179,132 +166,6 @@ fn next_search(search_on: &SearchOn) -> SearchOn {
         Title => Url,
         Url => Title,
     }
-}
-
-fn order_by_match(
-    history: &mut [HistoryData],
-    user_search: &mut String,
-    search_on: &SearchOn,
-) -> Vec<HistoryData> {
-    let mut matcher = Matcher::new(Config::DEFAULT);
-    let pattern = Pattern::new(
-        &*user_search,
-        CaseMatching::Ignore,
-        Normalization::Smart,
-        *SEARCH_TYPE,
-    );
-    let mut data_2_score = history
-        .iter()
-        .map(|h| {
-            let match_on = search_on_history(h, search_on);
-            (
-                h,
-                pattern.score(Utf32Str::new(match_on, &mut vec![]), &mut matcher),
-            )
-        })
-        .filter(|(_, score)| score.is_some())
-        .collect::<Vec<(&HistoryData, Option<u32>)>>();
-    data_2_score.sort_by(|(h1, a), (h2, b)| {
-        match a.unwrap_or_else(|| 0).cmp(&b.unwrap_or_else(|| 0)) {
-            Ordering::Less => Ordering::Less,
-            Ordering::Equal => h1.time.cmp(&h2.time),
-            Ordering::Greater => Ordering::Greater,
-        }
-    });
-    data_2_score.into_iter().map(|(a, _)| a.clone()).collect()
-}
-
-fn search_on_history<'a>(history: &'a HistoryData, search_on: &'a SearchOn) -> &'a str {
-    match search_on {
-        Title => &history.title,
-        Url => &history.url,
-    }
-}
-
-fn create_table<'a>(rows: &mut [Row<'a>]) -> Table<'a> {
-    let table = Table::from_iter(rows.to_owned())
-        .widths([
-            Constraint::Percentage(50),
-            Constraint::Percentage(40),
-            Constraint::Percentage(10),
-        ])
-        .column_spacing(1)
-        .highlight_symbol("> ")
-        .row_highlight_style(Style::default().add_modifier(Modifier::BOLD));
-    table
-}
-
-fn create_rows(
-    history: &[HistoryData],
-    user_search: &str,
-    search_on: &SearchOn,
-) -> Vec<Row<'static>> {
-    let rows: Vec<Row> = history
-        .iter()
-        .map(|h| match search_on {
-            Title => {
-                let cell = vec![
-                    Cell::from(highlight_text(clip_if_needed(&h.title, 100), user_search))
-                        .style(*TITLE_COLOR),
-                    Cell::from(clip_if_needed(&h.url, 60)).style(*URL_COLOR),
-                    Cell::from(date_to_display(&h.time)).style(*TIME_COLOR),
-                ];
-                Row::new(cell)
-            }
-            Url => {
-                let cells = vec![
-                    Cell::from(clip_if_needed(&h.title, 100)).style(*TITLE_COLOR),
-                    Cell::from(highlight_text(clip_if_needed(&h.url, 60), user_search))
-                        .style(*URL_COLOR),
-                    Cell::from(date_to_display(&h.time)).style(*TIME_COLOR),
-                ];
-                Row::new(cells)
-            }
-        })
-        .collect();
-    rows
-}
-
-fn highlight_text(plain_text: String, user_search: &str) -> Line<'static> {
-    let user_search = user_search.replace(' ', "");
-    if user_search.is_empty() || plain_text.is_empty() {
-        return Line::from(plain_text);
-    }
-    let mut matcher = Matcher::new(Config::DEFAULT);
-
-    let mut indices = vec![];
-    let mut binding1 = vec![];
-    let mut binding2 = vec![];
-    let haystack = Utf32Str::new(&plain_text, &mut binding1);
-    let lowercase = user_search.to_lowercase(); // Panics with uppercase??? Lowercase still matches
-    let needle = Utf32Str::new(&lowercase, &mut binding2);
-    matcher.fuzzy_indices(haystack, needle, &mut indices);
-    if indices.is_empty() {
-        return Line::from(plain_text);
-    }
-    let mut idx = indices.remove(0);
-    let mut current = String::new();
-    let mut spans = vec![];
-    let mut found = false;
-    for (c, i) in plain_text.chars().zip(0..) {
-        if found || i < idx {
-            current.push(c);
-        } else {
-            spans.push(Span::from(current.clone()));
-            current = String::new();
-            spans.push(Span::styled(
-                String::from(c),
-                Style::from(Color::Red).add_modifier(Modifier::BOLD),
-            ));
-            if indices.is_empty() {
-                found = true;
-            } else {
-                idx = indices.remove(0);
-            }
-        }
-    }
-    spans.push(Span::from(current));
-    Line::from(spans)
 }
 
 fn handle_input() -> Action {
@@ -346,37 +207,6 @@ pub enum SearchOn {
     Url,
 }
 
-fn clip_if_needed(text: &str, max_length: usize) -> String {
-    if text.len() > max_length {
-        return format!("{}...", &text[0..max_length - 3]);
-    }
-    text.to_string()
-}
-
-fn date_to_display(date: &NaiveDateTime) -> String {
-    let duration = Utc::now().signed_duration_since(date.and_utc());
-    if duration.num_weeks() > 0 {
-        return format_time(duration.num_weeks(), "weeks");
-    }
-    if duration.num_days() > 0 {
-        return format_time(duration.num_days(), "days");
-    }
-    if duration.num_hours() > 0 {
-        return format_time(duration.num_hours(), "hours");
-    }
-    if duration.num_minutes() > 0 {
-        return format_time(duration.num_minutes(), "minutes");
-    }
-    if duration.num_seconds() > 0 {
-        return format_time(duration.num_seconds(), "seconds");
-    }
-    "Date could not be displayed".to_string()
-}
-
-fn format_time(amount: i64, time_measurement: &str) -> String {
-    format!("{amount} {time_measurement} ago")
-}
-
 fn open_browser(history: &[HistoryData], state: &TableState) {
     let idx = state.selected().unwrap_or(0);
     history
@@ -391,35 +221,4 @@ fn open_browser(history: &[HistoryData], state: &TableState) {
                 false,
             );
         });
-}
-
-fn draw_search_text<'a>(user_input: &'a str, search_on: &'a SearchOn) -> Paragraph<'a> {
-    let searched_on_text = searched_on_to_string(search_on);
-    Paragraph::new(
-        Line::from(format!(" [{searched_on_text}] {user_input}"))
-            .style(TEXT_COLOR.add_modifier(Modifier::BOLD)),
-    )
-}
-fn draw_history_count(row_count: u16) -> ratatui::prelude::Text<'static> {
-    ratatui::prelude::Text::from(vec![
-        Line::default(), // Move to the bottom line
-        Line::from(count_result_text(row_count))
-            .style(crate::tui::display::TUI_BORDER_COLOR.add_modifier(Modifier::BOLD))
-            .alignment(Alignment::Right),
-    ])
-}
-
-fn count_result_text(row_count: u16) -> String {
-    if row_count == 1 {
-        format!("{row_count} result ")
-    } else {
-        format!("{row_count} results ")
-    }
-}
-
-fn searched_on_to_string(search_on: &SearchOn) -> String {
-    match search_on {
-        Title => String::from("TITLE"),
-        Url => String::from("URL"),
-    }
 }
