+++ src/tui/mod.rs
@@ -1,3 +1,7 @@
 pub mod browser;
+mod browser_widgets;
 mod display;
+mod general_widgets;
 pub mod history;
+mod history_search;
+mod history_widgets;
